Changes to the track view screen, fixed minor bugs, changed variables and deleted unnecessary button

+++ app/src/main/java/com/example/trackpro/TrackBuilderScreen.kt
@@ -281,7 +281,7 @@ fun TrackBuilderScreen(
                 } else {
                     coroutineScope.launch(Dispatchers.IO) {
                         stopBatchInsert()
-                        postProc.processTrackPoints(trackId = trackID.toInt())
+                        postProc.processTrackPoints(trackId = trackID)
                         trackID = -1
                     }
                 }
